Extract transaction collection component
